Use exit code for auth check
I started this change because run_process has a bug where it does not return stdout/stderr for some reason (which still very much affects us in other ways).

By cleaning up the auth code a bit I ended up with a solution that should be much more robust going forward. I might want to refactor this to contain the full request in liqudsoap but that is too much for this commit.

For now this gets source streaming back up and running for those that need it and want to be on the crazy bleeding edge.