Upgrade fullcalendar to 2.0 and move to bower
I had to upgrade fullcalendar to get a version that supprts the kind of
feature that where previously made to work with local patches. I tried
editing calendars in differenct timezones than locally and everything
seems to add up when we use the native "timezone" support that comes
with fullcalendar 2.x.